fix:scr.pyをルートドメインに移動
